backend/requirements.txt
accelerate==1.11.0
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.11.0
bitsandbytes==0.48.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.2
defusedxml==0.7.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.118.0
filelock==3.20.0
fsspec==2025.9.0
ftfy==6.3.1
greenlet==3.2.4
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
langdetect==1.0.9
lxml==6.0.2
MarkupSafe==3.0.3
ml_dtypes==0.5.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
onnx==1.19.1
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==12.0.0
protobuf==6.33.0
psutil==7.1.2
psycopg2==2.9.10
pyasn1==0.6.1
pyclipper==1.3.0.post6
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
PyPDF2==3.0.1
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-doctr==1.0.0
python-docx==1.2.0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2025.10.23
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.2
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
triton==3.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
validators==0.35.0
wcwidth==0.2.14
xlsxwriter==3.2.9
